ロジックも別ファイルに切り出す #2
