+++ src/app/components/buyButtonHandler/buyButtonHandler.ts
@@ -1,0 +1,4 @@
+export const handleBuyButton = (bookInfo: any) => {
+  const hanmotoLink = `https://www.hanmoto.com/bd/isbn/${bookInfo.isbn}`;
+  window.open(hanmotoLink);
+};
